Sistema de notificaciones
* Cuando un delivery se postula a un pedido. Al dueño del pedido le aparecerá una notificación.
* Al presionar sobre la notificación podrá acceder al detalle de la misma.

+++ TeLoBusco/RepositorioClases/Postulacion.cs
@@ -12,7 +12,7 @@ namespace RepositorioClases
         public int IdPedido { get; set; }
         public string IdUsuarioPostulado { get; set; }
         public int IdEstadoPostulacion { get; set; }
-        public float TiempoEstimado { get; set; }
+        public double TiempoEstimado { get; set; }
         public decimal Precio { get; set; }
     }
 }
